refactor: 커스텀 훅 네이밍 api와 일관되게 수정
